retrieval_service/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
faiss-cpu==1.13.2
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
kagglehub==0.4.0
kagglesdk==0.1.14
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
packaging==24.2
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
protobuf==6.33.2
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
